+++ src/imio/events/core/__init__.py
@@ -2,9 +2,4 @@
 
 from imio.smartweb.common.config import VOCABULARIES_MAPPING
 
-VOCABULARIES_MAPPING.update(
-    {
-        "category": "imio.events.vocabulary.EventsCategories",
-        "local_category": "imio.events.vocabulary.EventsLocalCategories",
-    }
-)
+VOCABULARIES_MAPPING.update({"category": "imio.events.vocabulary.EventsCategories"})