+++ src/imio/events/core/tests/test_event.py
@@ -172,7 +172,11 @@ class TestEvent(unittest.TestCase):
         vocabulary = factory(event)
         self.assertEqual(len(vocabulary), 0)
 
-        self.entity.local_categories = "First\nSecond\nThird"
+        self.entity.local_categories = [
+            {"fr": "First", "nl": "", "de": "", "en": ""},
+            {"fr": "Second", "nl": "", "de": "", "en": ""},
+            {"fr": "Third", "nl": "", "de": "", "en": ""},
+        ]
         vocabulary = factory(event)
         self.assertEqual(len(vocabulary), 3)